+++ web/src/main/java/com/mtn/controller/ShoppingCenterAccessController.java
@@ -2,6 +2,7 @@ package com.mtn.controller;
 
 import com.mtn.model.domain.ShoppingCenterAccess;
 import com.mtn.model.view.SimpleShoppingCenterAccessView;
+import com.mtn.service.SecurityService;
 import com.mtn.service.ShoppingCenterAccessService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -16,15 +17,21 @@ public class ShoppingCenterAccessController {
 
     @Autowired
     private ShoppingCenterAccessService accessService;
+    @Autowired
+    private SecurityService securityService;
 
     @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity deleteOne(@PathVariable("id") Integer id) {
+        securityService.checkPermission("SHOPPING_CENTER_SURVEYS_DELETE");
+
         accessService.deleteOne(id);
         return ResponseEntity.noContent().build();
     }
 
     @RequestMapping(path = "/{id}", method = RequestMethod.GET)
     public ResponseEntity findOne(@PathVariable("id") Integer id) {
+        securityService.checkPermission("SHOPPING_CENTER_SURVEYS_READ");
+
         ShoppingCenterAccess domainModel = accessService.findOne(id);
         if (domainModel != null) {
             return ResponseEntity.ok(new SimpleShoppingCenterAccessView(domainModel));
@@ -35,6 +42,8 @@ public class ShoppingCenterAccessController {
 
     @RequestMapping(path = "/{id}", method = RequestMethod.PUT)
     public ResponseEntity updateOne(@PathVariable("id") Integer id, @RequestBody ShoppingCenterAccess request) {
+        securityService.checkPermission("SHOPPING_CENTER_SURVEYS_UPDATE");
+
         ShoppingCenterAccess domainModel = accessService.updateOne(id, request);
         return ResponseEntity.ok(new SimpleShoppingCenterAccessView(domainModel));
     }
