+++ web/src/main/java/com/mtn/controller/ShoppingCenterTenantController.java
@@ -2,6 +2,7 @@ package com.mtn.controller;
 
 import com.mtn.model.domain.ShoppingCenterTenant;
 import com.mtn.model.view.SimpleShoppingCenterTenantView;
+import com.mtn.service.SecurityService;
 import com.mtn.service.ShoppingCenterTenantService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -16,15 +17,21 @@ public class ShoppingCenterTenantController {
 
     @Autowired
     private ShoppingCenterTenantService tenantService;
+    @Autowired
+    private SecurityService securityService;
 
     @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity deleteOne(@PathVariable("id") Integer id) {
+        securityService.checkPermission("SHOPPING_CENTER_SURVEYS_UPDATE");
+
         tenantService.deleteOne(id);
         return ResponseEntity.noContent().build();
     }
 
     @RequestMapping(path = "/{id}", method = RequestMethod.GET)
     public ResponseEntity findOne(@PathVariable("id") Integer id) {
+        securityService.checkPermission("SHOPPING_CENTER_SURVEYS_READ");
+
         ShoppingCenterTenant domainModel = tenantService.findOne(id);
         if (domainModel != null) {
             return ResponseEntity.ok(new SimpleShoppingCenterTenantView(domainModel));
@@ -35,6 +42,8 @@ public class ShoppingCenterTenantController {
 
     @RequestMapping(path = "/{id}", method = RequestMethod.PUT)
     public ResponseEntity updateOne(@PathVariable("id") Integer id, @RequestBody ShoppingCenterTenant request) {
+        securityService.checkPermission("SHOPPING_CENTER_SURVEYS_UPDATE");
+
         ShoppingCenterTenant domainModel = tenantService.updateOne(id, request);
         return ResponseEntity.ok(new SimpleShoppingCenterTenantView(domainModel));
     }
