+++ web/src/main/java/com/mtn/controller/StoreController.java
@@ -2,6 +2,7 @@ package com.mtn.controller;
 
 import com.mtn.model.domain.Store;
 import com.mtn.model.view.StoreView;
+import com.mtn.service.SecurityService;
 import com.mtn.service.StoreService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -16,15 +17,21 @@ public class StoreController {
 
     @Autowired
     private StoreService storeService;
+    @Autowired
+    private SecurityService securityService;
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity deleteOne(@PathVariable("id") Integer id) {
+        securityService.checkPermission("STORES_DELETE");
+
         storeService.deleteOne(id);
         return ResponseEntity.noContent().build();
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public ResponseEntity findOne(@PathVariable("id") Integer id) {
+        securityService.checkPermission("STORES_READ");
+
         Store domainModel = storeService.findOneUsingSpecs(id);
         if (domainModel != null) {
             return ResponseEntity.ok(new StoreView(domainModel));
@@ -35,6 +42,8 @@ public class StoreController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
     public ResponseEntity updateOne(@PathVariable("id") Integer id, @RequestBody StoreView request) {
+        securityService.checkPermission("STORES_UPDATE");
+
         Store domainModel = storeService.updateOne(id, new Store((StoreView) request)); //Cast ensures correct constructor is called
         return ResponseEntity.ok(new StoreView(domainModel));
     }
