+++ web/src/main/java/com/mtn/controller/UserProfileController.java
@@ -5,6 +5,7 @@ import com.mtn.model.domain.UserIdentity;
 import com.mtn.model.domain.UserProfile;
 import com.mtn.model.view.SimpleUserIdentityView;
 import com.mtn.model.view.UserProfileView;
+import com.mtn.service.SecurityService;
 import com.mtn.service.UserIdentityService;
 import com.mtn.service.UserProfileService;
 import org.apache.commons.lang3.StringUtils;
@@ -26,12 +27,15 @@ public class UserProfileController {
 
     @Autowired
     private UserIdentityService userIdentityService;
-
     @Autowired
     private UserProfileService userProfileService;
+    @Autowired
+    private SecurityService securityService;
 
     @RequestMapping(method = RequestMethod.POST)
     public ResponseEntity addOne(@RequestBody UserProfile request) {
+        securityService.checkPermission("USERS_CREATE");
+
         UserProfile domainModel = userProfileService.addOne(request);
         return ResponseEntity.ok(new UserProfileView(domainModel));
     }
@@ -44,12 +48,16 @@ public class UserProfileController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity deleteOne(@PathVariable("id") Integer id) {
+        securityService.checkPermission("USERS_DELETE");
+
         userProfileService.deleteOne(id);
         return ResponseEntity.noContent().build();
     }
 
     @RequestMapping(method = RequestMethod.GET)
     public ResponseEntity findAll(@RequestParam(value = "q", required = false) String q, Pageable page) {
+        securityService.checkPermission("USERS_READ");
+
         Page<UserProfile> domainModels;
         if (StringUtils.isNotBlank(q)) {
             domainModels = userProfileService.query(q, page);
@@ -68,6 +76,8 @@ public class UserProfileController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public ResponseEntity findOne(@PathVariable("id") String id) {
+        securityService.checkPermission("USERS_READ");
+
         UserProfile domainModel;
         if (StringUtils.isNumeric(id)) {
             domainModel = userProfileService.findOneUsingSpecs(Integer.parseInt(id));
@@ -84,6 +94,8 @@ public class UserProfileController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
     public ResponseEntity updateOne(@PathVariable("id") Integer id, @RequestBody UserProfile request) {
+        securityService.checkPermission("USERS_UPDATE");
+
         UserProfile domainModel = userProfileService.updateOne(id, request);
         return ResponseEntity.ok(new UserProfileView(domainModel));
     }
